requirements.txt
torchvision==0.12.0
pytorch-ignite==0.4.9
hydra-core==1.1.1
omegaconf==2.1.1
scikit-learn==1.1.0
numpy==1.22.3
pandas==1.4.2
matplotlib==3.5.2
seaborn==0.11.2
scikit-image==0.19.2
h5py==3.6.0
attrdict==2.0.1
tqdm~=4.64
deepdiff==5.8.1
PyYAML==6.0
Pillow==9.1.0
protobuf==3.20.3
black==22.3.0
isort==5.10.1
deepspeed
accelerate==0.16.0
deepdiff
tensorflow==2.12
dominate
func-timeout
omegaconf
h5py
matplotlib
tensorboard==2.12
torchmetrics==0.11.4
transformers==4.28.1
einops
ipdb
pycocotools
scikit-image
wandb
